Precision test

Tests the null hypothesis of precision fixed in RBS models against the alternative of precision variable.

grad_test_bp(modelh0,modelh1)

grad_test_ga(modelh0, modelh1)

grad_test_ig(modelh0, modelh1)

grad_test_wei(modelh0, modelh1)

dRBS(x, mu = 1, sigma = 1, log = FALSE)

score_test_bp(modelh0, modelh1)

score_test_rbs(modelh0, modelh1)

score_test_ga(modelh0, modelh1)

score_test_ig(modelh0, modelh1)

score_test_wei(modelh0, modelh1)

wald_test(modelh1)

Arguments

modelh0
model under null hypothesis.

modelh1
model under alternative hypothesis.

x
vector of quantiles.

mu, sigma
the (positive) location and precision parameter.

log
logical; The logarithm of the density is returned if the value is TRUE.

Value

A list with class "htest" containing the following components:

statistic the value of the test statistic.

parameter the degrees of freedom for the test statistic.

p.value the p-value for the test.

method a character string indicating what type of likelihood ratio test was performed.

data.name a character string giving the name(s) of the data

Author

Manoel Santos-Neto manoelferreira@uaest.ufcg.edu.br